Move shower hit and cluster classes from THcShower.h header file into THcShowerHitClass.h file.
